Reorder go packages & consolidate db interactions
All go files exists in src/ladder, although only player.go is in use in
the main.

Introduced insert.go and select.go which insert a full player into the
database and select all players from the database. A new function
SelectOnSomething could be introduced and would probably be helpful for
getting particular players. aga_ids should be unique, although this
isn't in the schema yet.